myBot/telegramDataAnalysis/requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
regex==2024.11.6
rsa==4.9
setuptools==75.1.0
Telethon==1.38.1
wheel==0.44.0
